+++ src/main/java/controller/MenuOption.java
@@ -1,4 +1,4 @@
-package model;
+package controller;
 
 public enum MenuOption {
